+++ @navikt/core/react/src/date/datepicker/DatePicker.tests.stories.tsx
@@ -13,6 +13,8 @@ export default {
   parameters: {
     chromatic: { disable: true },
   },
+  /* TODO: Temp disabled CI-tests */
+  tags: ["!play-fn", "skip-test"],
 } satisfies Meta<typeof DatePicker>;
 
 type Story = StoryObj<typeof DatePicker>;