+++ @navikt/core/react/src/tabs/Tabs.stories.tsx
@@ -3,6 +3,7 @@ import React, { useState } from "react";
 import { DishwasherIcon, FreezerIcon, MugIcon } from "@navikt/aksel-icons";
 import { Tabs } from ".";
 import { VStack } from "../layout/stack";
+import { renderStoriesForChromatic } from "../util/renderStoriesForChromatic";
 
 export default {
   title: "ds-react/Tabs",
@@ -217,24 +218,25 @@ export const Overflow = () => (
 );
 
 export const Fill = () => (
-  <Tabs defaultValue="test2" fill>
-    <Tabs.List>
-      <Tabs.Tab value="test1" icon={<MugIcon aria-hidden />} label="Skap" />
-      <Tabs.Tab
-        value="test2"
-        icon={<DishwasherIcon aria-hidden />}
-        label="Oppvaskmaskin"
-      />
-      <Tabs.Tab
-        value="test3"
-        icon={<FreezerIcon aria-hidden />}
-        label="Fryser"
-      />
-    </Tabs.List>
-    <Panel />
-  </Tabs>
+  <div style={{ minWidth: 500 }}>
+    <Tabs defaultValue="test2" fill>
+      <Tabs.List>
+        <Tabs.Tab value="test1" icon={<MugIcon aria-hidden />} label="Skap" />
+        <Tabs.Tab
+          value="test2"
+          icon={<DishwasherIcon aria-hidden />}
+          label="Oppvaskmaskin"
+        />
+        <Tabs.Tab
+          value="test3"
+          icon={<FreezerIcon aria-hidden />}
+          label="Fryser"
+        />
+      </Tabs.List>
+      <Panel />
+    </Tabs>
+  </div>
 );
-
 Fill.parameters = {
   layout: "fullscreen",
 };
@@ -280,42 +282,15 @@ export const ColorRole = () => (
   </div>
 );
 
-export const Chromatic = {
-  render: () => (
-    <VStack gap="6" align="center">
-      <div>
-        <h2>Small</h2>
-        <Small />
-      </div>
-      <div>
-        <h2>Controlled</h2>
-        <Controlled />
-      </div>
-      <div>
-        <h2>IconPosition</h2>
-        <IconPosition />
-      </div>
-      <div>
-        <h2>Icon</h2>
-        <Icon />
-      </div>
-      <div>
-        <h2>Overflow</h2>
-        <Overflow />
-      </div>
-      <div>
-        <h2>Fill</h2>
-        <div style={{ minWidth: 600 }}>
-          <Fill />
-        </div>
-      </div>
-      <div>
-        <h2>ColorRole</h2>
-        <ColorRole />
-      </div>
-    </VStack>
-  ),
-  parameters: {
-    chromatic: { disable: false, delay: 300 },
-  },
+export const Chromatic = renderStoriesForChromatic({
+  Small,
+  Controlled,
+  IconPosition,
+  Icon,
+  Overflow,
+  Fill,
+  ColorRole,
+});
+Chromatic.parameters = {
+  chromatic: { disable: false, delay: 300 },
 };