+++ @navikt/core/react/src/typography/stories/bodyshort.stories.tsx
@@ -2,6 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import React from "react";
 import { expect, within } from "storybook/test";
 import { VStack } from "../../layout/stack";
+import { renderStoriesForChromatic } from "../../util/renderStoriesForChromatic";
 import { BodyShort } from "../BodyShort";
 
 const meta: Meta<typeof BodyShort> = {
@@ -171,46 +172,14 @@ export const ColorRole: Story = {
   ),
 };
 
-export const Chromatic: Story = {
-  render: (...props) => (
-    <div>
-      <div>
-        <h2>Large</h2>
-        <h3>Size</h3>
-        {SizeLarge.render?.(...props)}
-        <h3>Spacing</h3>
-        {SpacingLarge.render?.(...props)}
-      </div>
-      <div>
-        <h2>Medium</h2>
-        <h3>Size</h3>
-        {SizeMedium.render?.(...props)}
-        <h3>Spacing</h3>
-        {SpacingMedium.render?.(...props)}
-      </div>
-      <div>
-        <h2>Small</h2>
-        <h3>Size</h3>
-        {SizeSmall.render?.(...props)}
-        <h3>Spacing</h3>
-        {SpacingSmall.render?.(...props)}
-      </div>
-
-      <div>
-        <h2>Align</h2>
-        {Align.render?.(...props)}
-      </div>
-      <div>
-        <h2>Override Tag</h2>
-        {OverrideTag.render?.(...props)}
-      </div>
-      <div>
-        <h2>ColorRole</h2>
-        {ColorRole.render?.(...props)}
-      </div>
-    </div>
-  ),
-  parameters: {
-    chromatic: { disable: false },
-  },
-};
+export const Chromatic = renderStoriesForChromatic({
+  SizeLarge,
+  SpacingLarge,
+  SizeMedium,
+  SpacingMedium,
+  SizeSmall,
+  SpacingSmall,
+  Align,
+  OverrideTag,
+  ColorRole,
+});