+++ aksel.nav.no/website/app/(routes)/(designsystemet)/grunnleggende/darkside/design-tokens/_ui/toolbar/SearchField.tsx
@@ -31,7 +31,6 @@ const SearchField = () => {
         name="tokenSearch"
         hideLabel
         variant="simple"
-        htmlSize="26"
         autoComplete="off"
       />
     </form>