+++ aksel.nav.no/website/components/website-modules/examples/withDsExample.tsx
@@ -1,8 +1,16 @@
+"use client";
+
 import cl from "clsx";
 import { useTheme } from "next-themes";
 import Head from "next/head";
 import { usePathname } from "next/navigation";
-import { type CSSProperties, ComponentType, useSyncExternalStore } from "react";
+import {
+  type CSSProperties,
+  ComponentType,
+  useEffect,
+  useState,
+  useSyncExternalStore,
+} from "react";
 import {
   LaptopIcon,
   MobileIcon,
@@ -23,6 +31,7 @@ type withDsT = {
    */
   variant?: "full" | "static" | "static-full" | "fullscreen";
   background?: "inverted" | "subtle";
+  minHeight?: string;
   showBreakpoints?: boolean;
   /**
    * Hides theme switch, makes sure to not use `AkselTheme` wrapper and forces light-mode.
@@ -32,7 +41,13 @@ type withDsT = {
 
 export const withDsExample = (
   Component: ComponentType,
-  { variant, background, showBreakpoints, legacyOnly = false }: withDsT = {},
+  {
+    variant,
+    background,
+    minHeight,
+    showBreakpoints,
+    legacyOnly = false,
+  }: withDsT = {},
 ) => {
   const DsHOC = (props: any) => {
     const { theme } = useTheme();
@@ -51,7 +66,7 @@ export const withDsExample = (
           [styles.containerStaticFull]: variant === "static-full",
           [styles.containerFullscreen]: variant === "fullscreen",
         })}
-        style={{ background: getBg(background) }}
+        style={{ background: getBg(background), minHeight }}
       >
         <Head>
           <title>
@@ -90,6 +105,16 @@ function ThemeWrapper({
   style: CSSProperties;
   className: string;
 }) {
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  if (!mounted) {
+    return null;
+  }
+
   if (useDarkside) {
     return (
       <AkselTheme hasBackground={false} className={className} asChild>
@@ -110,7 +135,7 @@ function getBg(background: withDsT["background"]): string {
     case "inverted":
       return "var(--ax-neutral-1000)";
     case "subtle":
-      return "var(--a-bg-neutral-soft)";
+      return "var(--ax-bg-neutral-soft)";
 
     default:
       return "var(--ax-bg-default)";