+++ aksel.nav.no/website/pages/eksempler/combobox/multi-select.tsx
@@ -27,7 +27,10 @@ const options = [
 ];
 
 // EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
-export default withDsExample(Example, { variant: "static" });
+export default withDsExample(Example, {
+  variant: "static",
+  minHeight: "300px",
+});
 
 /* Storybook story */
 export const Demo = {