+++ aksel.nav.no/website/sanity/sanity.config.ts
@@ -69,7 +69,7 @@ export const workspaceConfig = defineConfig([
       }),
     ],
     releases: {
-      enabled: false,
+      enabled: true,
     },
   },
   {